Redux

+++ functions/install_minecraft.ps1
@@ -10,11 +10,15 @@ Function Get-MCBRBinaries {
         
     #################### MineCraftBedrock Install ################
     Get-MCBRWebrequest 
-    Write-Host '****   Downloading  Minecraft Bedrock  ****' -F M -B Black
+    $global:package = 'Minecraft Bedrock'
+    $global:infomessage = "Downloading"
+    Get-Infomessage
     Invoke-WebRequest -uri $mcbrWebResponse.href -O bedrock-server.zip
-    Write-Host '****   Extracting  Minecraft Bedrock   ****' -F M -B Black
+    $global:infomessage = "Extracting"
+    Get-Infomessage
     Expand-Archive "bedrock-server.zip" "bedrock-server" -Force -ea SilentlyContinue
     Move-Item bedrock-server\* $serverfiles\ -Force -ea SilentlyContinue
+    New-Item $serverfiles\version.txt -Force 3>&1 2>&1 > $ssmlog
     Add-Content $serverfiles\version.txt $mcbrWebResponse.href -Force
     Remove-Item bedrock-server -Recurse -Force -ea SilentlyContinue
     ##############################################################
@@ -24,17 +28,22 @@ Function Get-MCjavaBinaries {
     #################### MineCraft Java Install ################
     Get-MCWebrequest 
     $mcWebResponse = ((Invoke-WebRequest "https://www.minecraft.net/en-us/download/server" -UseBasicParsing ).Links | Where-Object { $_.href -like "https://launcher.mojang.com/v1/objects/*/server.jar" })
-    Write-Host '****   Downloading  Minecraft Java   ****' -F M -B Black  
+    # Write-Host '****   Downloading  Minecraft Java   ****' -F M -B Black
+    $global:package = 'Minecraft Java'
+    $global:infomessage = "Downloading"  
     Invoke-WebRequest -uri $mcWebResponse.href -O server.jar 
     # $mcWebResponse.outerText
     # Expand-Archive "bedrock-server.zip" "bedrock-server" -Force -ea SilentlyContinue
     Move-Item server.jar $serverfiles\ -Force -ea SilentlyContinue
+    New-Item $serverfiles\version.txt -Force 3>&1 2>&1 > $ssmlog
     Add-Content $serverfiles\version.txt $mcvWebResponse -Force
     Set-Location $serverdir
-    Write-Host '****   Starting server to create and edit eula   ****' -F M -B Black
-    Start-Process CMD "/c start java -Xms1024M -Xmx1024M -jar server.jar nogui" -Wait
-    Start-Sleep 7
-    ((Get-Content -path eula.txt -Raw) -replace "false", "true") | Set-Content -Path eula.txt
+    # Write-Host '****   Starting server to create and edit eula   ****' -F M -B Black
+    If (!(Test-Path eula.txt )) {
+        Start-Process CMD "/c start java -Xms1024M -Xmx1024M -jar server.jar nogui" -Wait
+        Start-Sleep 3
+        ((Get-Content -path eula.txt -Raw) -replace "false", "true") | Set-Content -Path eula.txt
+    }
     Set-Location $currentdir
     # Add-Content $serverfiles\eula.txt 'eula=true' -Force
     # Remove-Item bedrock-server -Recurse -Force -ea SilentlyContinue
